requirements.txt
# Core geospatial
geopandas>=0.12
rasterio>=1.3
shapely>=1.8
pyproj>=3.4
fiona>=1.8

# OSM data
osmnx>=1.3
overpy>=0.6

# Kriging
pykrige>=1.6
scikit-learn>=1.2

# Graph Neural Networks
torch>=2.0
torch-geometric>=2.3

# Spatial statistics
libpysal>=4.7
esda>=2.4
pointpats>=2.3

# Visualization & dashboard
folium>=0.14
matplotlib>=3.7
streamlit>=1.28
streamlit-folium>=0.15
plotly>=5.15
branca>=0.6

# Pipeline utilities
click>=8.1
pyyaml>=6.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
tqdm>=4.65
loguru>=0.7
requests>=2.28

# AlphaEarth / Earth Engine
earthengine-api>=0.1.370
umap-learn>=0.5

# Image processing (raster overlay)
Pillow>=9.0

# Export
fpdf2>=2.7
